Move the test helpers into a scheme folder
