+++ test/support/helpers/water/billing-batch.helper.js
@@ -1,33 +1,38 @@
 'use strict'
 
 /**
- * @module LicenceHelper
+ * @module BillingBatchHelper
  */
 
-const LicenceModel = require('../../../app/models/licence.model.js')
+const BillingBatchModel = require('../../../../app/models/water/billing-batch.model.js')
 
 /**
- * Add a new licence
+ * Add a new billing batch
  *
  * If no `data` is provided, default values will be used. These are
  *
- * - `licenceRef` - 01/123
  * - `regionId` - bd114474-790f-4470-8ba4-7b0cc9c225d7
+ * - `batchType` - supplementary
+ * - `fromFinancialYearEnding` - 2023
+ * - `toFinancialYearEnding` - 2023
+ * - `status` - processing
+ * - `scheme` - sroc
+ * - `source` - wrls
  *
  * @param {Object} [data] Any data you want to use instead of the defaults used here or in the database
  *
- * @returns {module:LicenceModel} The instance of the newly created record
+ * @returns {module:BillingBatchModel} The instance of the newly created record
  */
-async function add (data = {}) {
+function add (data = {}) {
   const insertData = defaults(data)
 
-  return LicenceModel.query()
+  return BillingBatchModel.query()
     .insert({ ...insertData })
     .returning('*')
 }
 
 /**
- * Returns the defaults used when creating a new licence
+ * Returns the defaults used when creating a new billing batch
  *
  * It will override or append to them any data provided. Mainly used by the `add()` method, we make it available
  * for use in tests to avoid having to duplicate values.
@@ -36,8 +41,13 @@ async function add (data = {}) {
  */
 function defaults (data = {}) {
   const defaults = {
-    licenceRef: '01/123',
-    regionId: 'bd114474-790f-4470-8ba4-7b0cc9c225d7'
+    regionId: 'bd114474-790f-4470-8ba4-7b0cc9c225d7',
+    batchType: 'supplementary',
+    fromFinancialYearEnding: 2023,
+    toFinancialYearEnding: 2023,
+    status: 'processing',
+    scheme: 'sroc',
+    source: 'wrls'
   }
 
   return {
