+++ test/support/helpers/water/licence.helper.js
@@ -1,38 +1,33 @@
 'use strict'
 
 /**
- * @module BillingBatchHelper
+ * @module LicenceHelper
  */
 
-const BillingBatchModel = require('../../../app/models/billing-batch.model.js')
+const LicenceModel = require('../../../../app/models/water/licence.model.js')
 
 /**
- * Add a new billing batch
+ * Add a new licence
  *
  * If no `data` is provided, default values will be used. These are
  *
+ * - `licenceRef` - 01/123
  * - `regionId` - bd114474-790f-4470-8ba4-7b0cc9c225d7
- * - `batchType` - supplementary
- * - `fromFinancialYearEnding` - 2023
- * - `toFinancialYearEnding` - 2023
- * - `status` - processing
- * - `scheme` - sroc
- * - `source` - wrls
  *
  * @param {Object} [data] Any data you want to use instead of the defaults used here or in the database
  *
- * @returns {module:BillingBatchModel} The instance of the newly created record
+ * @returns {module:LicenceModel} The instance of the newly created record
  */
-function add (data = {}) {
+async function add (data = {}) {
   const insertData = defaults(data)
 
-  return BillingBatchModel.query()
+  return LicenceModel.query()
     .insert({ ...insertData })
     .returning('*')
 }
 
 /**
- * Returns the defaults used when creating a new billing batch
+ * Returns the defaults used when creating a new licence
  *
  * It will override or append to them any data provided. Mainly used by the `add()` method, we make it available
  * for use in tests to avoid having to duplicate values.
@@ -41,13 +36,8 @@ function add (data = {}) {
  */
 function defaults (data = {}) {
   const defaults = {
-    regionId: 'bd114474-790f-4470-8ba4-7b0cc9c225d7',
-    batchType: 'supplementary',
-    fromFinancialYearEnding: 2023,
-    toFinancialYearEnding: 2023,
-    status: 'processing',
-    scheme: 'sroc',
-    source: 'wrls'
+    licenceRef: '01/123',
+    regionId: 'bd114474-790f-4470-8ba4-7b0cc9c225d7'
   }
 
   return {
